bazel: break out more libraries into their respective directories.
Added
  * src/ant/BUILD
  * src/cts/BUILD
  * src/dpl/BUILD
  * src/dbSta/BUILD
  * src/grt/BUILD
  * src/rsz/BUILD
  * src/stt/BUILD
