+++ MODULE.bazel
@@ -34,12 +34,15 @@ bazel_dep(name = "boost.asio", version = BOOST_VERSION)
 bazel_dep(name = "boost.beast", version = BOOST_VERSION + ".bcr.1")
 bazel_dep(name = "boost.bind", version = BOOST_VERSION)
 bazel_dep(name = "boost.config", version = BOOST_VERSION)
+bazel_dep(name = "boost.container_hash", version = BOOST_VERSION)
 bazel_dep(name = "boost.core", version = BOOST_VERSION)
+bazel_dep(name = "boost.format", version = BOOST_VERSION + ".bcr.2")
 bazel_dep(name = "boost.fusion", version = BOOST_VERSION)
 bazel_dep(name = "boost.geometry", version = BOOST_VERSION + ".bcr.2")
 bazel_dep(name = "boost.graph", version = BOOST_VERSION + ".bcr.2")
 bazel_dep(name = "boost.heap", version = BOOST_VERSION)
 bazel_dep(name = "boost.icl", version = BOOST_VERSION)
+bazel_dep(name = "boost.iterator", version = BOOST_VERSION)
 bazel_dep(name = "boost.json", version = BOOST_VERSION + ".bcr.2")
 bazel_dep(name = "boost.lambda", version = BOOST_VERSION + ".bcr.2")
 bazel_dep(name = "boost.multi_array", version = BOOST_VERSION + ".bcr.2")
@@ -47,10 +50,14 @@ bazel_dep(name = "boost.optional", version = BOOST_VERSION)
 bazel_dep(name = "boost.phoenix", version = BOOST_VERSION)
 bazel_dep(name = "boost.polygon", version = BOOST_VERSION + ".bcr.2")
 bazel_dep(name = "boost.property_tree", version = BOOST_VERSION + ".bcr.2")
+bazel_dep(name = "boost.random", version = BOOST_VERSION + ".bcr.1")
 bazel_dep(name = "boost.regex", version = BOOST_VERSION)
 bazel_dep(name = "boost.spirit", version = BOOST_VERSION)
 bazel_dep(name = "boost.stacktrace", version = BOOST_VERSION)
 bazel_dep(name = "boost.thread", version = BOOST_VERSION)
+bazel_dep(name = "boost.tokenizer", version = BOOST_VERSION)
+bazel_dep(name = "boost.unordered", version = BOOST_VERSION)
+bazel_dep(name = "boost.utility", version = BOOST_VERSION)
 bazel_dep(name = "cudd", version = "3.0.0")
 bazel_dep(name = "eigen", version = "3.4.0.bcr.2")
 bazel_dep(name = "or-tools", version = "9.12")
