+++ src/ant/include/ant/AntennaChecker.hh
@@ -8,6 +8,7 @@
 #include <queue>
 #include <set>
 
+#include "ant/GlobalRouteSource.hh"
 #include "odb/db.h"
 #include "odb/dbWireGraph.h"
 #include "utl/Logger.h"
@@ -89,16 +90,6 @@ struct ViolationReport
   ViolationReport() { violated = false; }
 };
 
-class GlobalRouteSource
-{
- public:
-  virtual ~GlobalRouteSource() = default;
-
-  virtual bool haveRoutes() = 0;
-  virtual void makeNetWires() = 0;
-  virtual void destroyNetWires() = 0;
-};
-
 struct Violation
 {
   int routing_level;
@@ -122,7 +113,7 @@ class AntennaChecker
   ~AntennaChecker();
 
   void init(odb::dbDatabase* db,
-            GlobalRouteSource* global_route_source,
+            ant::GlobalRouteSource* global_route_source,
             utl::Logger* logger);
 
   // net nullptr -> check all nets
@@ -210,7 +201,7 @@ class AntennaChecker
 
   odb::dbDatabase* db_{nullptr};
   odb::dbBlock* block_{nullptr};
-  GlobalRouteSource* global_route_source_{nullptr};
+  ant::GlobalRouteSource* global_route_source_{nullptr};
   utl::Logger* logger_{nullptr};
   std::map<odb::dbTechLayer*, AntennaModel> layer_info_;
   int net_violation_count_{0};
