+++ src/grt/src/MakeWireParasitics.cpp
@@ -9,9 +9,11 @@
 #include <utility>
 #include <vector>
 
+#include "Net.h"
 #include "db_sta/SpefWriter.hh"
 #include "db_sta/dbNetwork.hh"
 #include "db_sta/dbSta.hh"
+#include "grt/GlobalRouter.h"
 #include "rsz/Resizer.hh"
 #include "sta/ArcDelayCalc.hh"
 #include "sta/Corner.hh"
