+++ src/grt/src/MakeWireParasitics.h
@@ -8,11 +8,9 @@
 #include <vector>
 
 #include "AbstractMakeWireParasitics.h"
-#include "FastRoute.h"
-#include "Grid.h"
-#include "Net.h"
+#include "Pin.h"
 #include "db_sta/dbSta.hh"
-#include "grt/GlobalRouter.h"
+#include "grt/RoutePt.h"
 #include "sta/Clock.hh"
 #include "sta/Set.hh"
 
@@ -44,7 +42,7 @@ class Resizer;
 }
 
 namespace grt {
-
+class GlobalRouter;
 class MakeWireParasitics : public AbstractMakeWireParasitics
 {
  public:
