+++ src/grt/src/fastroute/src/FastRoute.cpp
@@ -13,6 +13,7 @@
 
 #include "AbstractFastRouteRenderer.h"
 #include "DataType.h"
+#include "grt/GRoute.h"
 #include "odb/db.h"
 #include "utl/Logger.h"
